refactor(upload): add canvas.js module to sepate logic
